+++ tests/_support/WpunitTesterConfig.php
@@ -1,9 +1,11 @@
 <?php
 /**
  * Define any constants not supported by WPLoader, such as the environment type.
- * 
+ *
  * See: https://github.com/lucatume/wp-browser/blob/master/docs/v3/modules/WPLoader.md,
  * parameter `configFile`.
+ *
+ * @package     ConvertKit
  */
 
-define( 'WP_ENVIRONMENT_TYPE', 'local' );
+define( 'WP_ENVIRONMENT_TYPE', $_ENV['WP_ENVIRONMENT_TYPE'] );
