+++ tests/wpunit/APITest.php
@@ -503,24 +503,19 @@ class APITest extends \Codeception\TestCase\WPTestCase
 
 	/**
 	 * Test that making a call with an expired access token results in refresh_token()
-	 * being automatically called to attempt to fetch a new access token.
+	 * not being automatically called, when the WordPress site isn't a production site.
 	 *
 	 * @since   2.0.2
 	 *
 	 * @return void
 	 */
-	public function testRefreshTokenWhenAccessTokenExpiredErrorOnProductionSite()
+	public function testRefreshTokenWhenAccessTokenExpiredErrorOnNonProductionSite()
 	{
-		// Confirm the refresh token action in the libraries is triggered when calling get_account().
+		// If the refresh token action in the libraries is triggered when calling get_account(), the test failed.
 		add_action(
 			'convertkit_api_refresh_token',
-			function(  $result ) {
-
-				$this->assertEquals( $result['access_token'], 'new-' . $_ENV['CONVERTKIT_OAUTH_ACCESS_TOKEN'] );
-				$this->assertEquals( $result['refresh_token'], 'new-' . $_ENV['CONVERTKIT_OAUTH_REFRESH_TOKEN'] );
-
-				// Mark test as passed.
-				$this->passed = true;
+			function() {
+				$this->fail('`convertkit_api_refresh_token` was triggered when calling `get_account` with an expired access token on a non-production site.');
 			}
 		);
 
@@ -530,9 +525,6 @@ class APITest extends \Codeception\TestCase\WPTestCase
 
 		// Run request, which will trigger the above filters as if the token expired and refreshes automatically.
 		$result = $this->api->get_account();
-
-		// Check that the refresh action ran; if it did, `passed` will be set to true.
-		$this->assertTrue($this->passed);
 	}
 
 	/**
@@ -6274,6 +6266,122 @@ class APITest extends \Codeception\TestCase\WPTestCase
 		$this->assertArrayHasKey('start_cursor', $pagination);
 		$this->assertArrayHasKey('end_cursor', $pagination);
 		$this->assertArrayHasKey('per_page', $pagination);
+	}
+
+	/**
+	 * Forces WordPress' wp_remote_*() functions to return a specific HTTP response code
+	 * and message by short circuiting using the `pre_http_request` filter.
+	 *
+	 * This emulates server responses that the API class has to handle from ConvertKit's API,
+	 * which we cannot easily recreate e.g. 500 or 502 errors.
+	 *
+	 * @since   1.0.0
+	 *
+	 * @param   int         $httpCode       HTTP Code.
+	 * @param   string      $httpMessage    HTTP Message.
+	 * @param   null|string $body           Response body.
+	 */
+	public function mockResponses( $httpCode, $httpMessage, $body = null )
+	{
+		add_filter(
+			'pre_http_request',
+			function( $response ) use ( $httpCode, $httpMessage, $body ) { // phpcs:ignore Generic.CodeAnalysis.UnusedFunctionParameter
+				return array(
+					'headers'       => array(),
+					'body'          => $body,
+					'response'      => array(
+						'code'    => $httpCode,
+						'message' => $httpMessage,
+					),
+					'cookies'       => array(),
+					'http_response' => null,
+				);
+			}
+		);
+	}
+
+	/**
+	 * Mocks an API response as if the Access Token expired.
+	 *
+	 * @since   2.0.2
+	 *
+	 * @param   mixed  $response       HTTP Response.
+	 * @param   array  $parsed_args    Request arguments.
+	 * @param   string $url            Request URL.
+	 * @return  mixed
+	 */
+	public function mockAccessTokenExpiredResponse( $response, $parsed_args, $url )
+	{
+		// Only mock requests made to the /account endpoint.
+		if ( strpos( $url, 'https://api.convertkit.com/v4/account' ) === false ) {
+			return $response;
+		}
+
+		// Remove this filter, so we don't end up in a loop when retrying the request.
+		remove_filter( 'pre_http_request', array( $this, 'mockAccessTokenExpiredResponse' ) );
+
+		// Return a 401 unauthorized response with the errors body as if the API
+		// returned "The access token expired".
+		return array(
+			'headers'       => array(),
+			'body'          => wp_json_encode(
+				array(
+					'errors' => array(
+						'The access token expired',
+					),
+				)
+			),
+			'response'      => array(
+				'code'    => 401,
+				'message' => 'The access token expired',
+			),
+			'cookies'       => array(),
+			'http_response' => null,
+		);
+	}
+
+	/**
+	 * Mocks an API response as if a refresh token was used to fetch new tokens.
+	 *
+	 * @since   2.0.2
+	 *
+	 * @param   mixed  $response       HTTP Response.
+	 * @param   array  $parsed_args    Request arguments.
+	 * @param   string $url            Request URL.
+	 * @return  mixed
+	 */
+	public function mockRefreshTokenResponse( $response, $parsed_args, $url )
+	{
+		// Only mock requests made to the /token endpoint.
+		if ( strpos( $url, 'https://api.convertkit.com/oauth/token' ) === false ) {
+			return $response;
+		}
+
+		// Remove this filter, so we don't end up in a loop when retrying the request.
+		remove_filter( 'pre_http_request', array( $this, 'mockRefreshTokenResponse' ) );
+
+		// Return a mock access and refresh token for this API request, as calling
+		// refresh_token results in a new access and refresh token being provided,
+		// which would result in other tests breaking due to changed tokens.
+		return array(
+			'headers'       => array(),
+			'body'          => wp_json_encode(
+				array(
+					'access_token'  => 'new-' . $_ENV['CONVERTKIT_OAUTH_ACCESS_TOKEN'],
+					'refresh_token' => 'new-' . $_ENV['CONVERTKIT_OAUTH_REFRESH_TOKEN'],
+					'token_type'    => 'bearer',
+					'created_at'    => strtotime( 'now' ),
+					'expires_in'    => 10000,
+					'scope'         => 'public',
+				)
+			),
+			'response'      => array(
+				'code'    => 200,
+				'message' => 'OK',
+			),
+			'cookies'       => array(),
+			'http_response' => null,
+		);
 	}
 
 	/**
